Rename multicurrency* files to multi_currency*
