+++ zrml/asset-router/src/pallet_impl/mod.rs
@@ -21,7 +21,7 @@ pub mod inspect;
 pub mod managed_destroy;
 pub mod multi_lockable_currency;
 pub mod multi_reserveable_currency;
-pub mod multicurrency;
-pub mod multicurrency_extended;
+pub mod multi_currency;
+pub mod multi_currency_extended;
 pub mod named_multi_reserveable_currency;
 pub mod transfer_all;
